requirements.txt
beautifulsoup4==4.9.0
click==7.1.2
googletrans==2.4.0
gTTS==2.1.1
gTTS-token==1.1.3
pyaml==20.4.0
pymongo==3.10.1
pyTelegramBotAPI==3.6.7
PyYAML==5.3.1
requests==2.7.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
